requirements.txt
# 核心框架
fastapi==0.115.0
uvicorn[standard]==0.24.0
pydantic>=2.9.0

# 阿里云百炼
dashscope==1.25.15

# 音频处理（pyaudio 仅本地播放用，容器内走 websocket 模式不需要）
pyyaml==6.0.1

# 工具
loguru==0.7.2
python-multipart==0.0.6
boto3
psycopg2-binary
requests
websockets
httpx

# RAG相关
chromadb>=0.6.0
sentence-transformers>=2.7.0
torch>=1.13.0
numpy>=1.21.0,<2.0.0

# LangChain相关
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-huggingface>=0.3.0

# Transformers（兼容版本）
transformers>=4.49.0